Change: 記事に関する機能を`/pkg/articles/`に配置するよう変更
main.goが煩雑になってきたので、記事を操作する処理をarticlesパッケージに切り出した。
`/pkg/`を作成したのはgoのスタンダードディレクトリ構成を参考にしてのこと。

参考: https://github.com/golang-standards/project-layout
